requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.14
charset-normalizer==2.1.1
discord.py==2.0.1
frozenlist==1.3.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
loguru==0.6.0
multidict==6.0.2
rfc3986==1.5.0
sniffio==1.3.0
yarl==1.8.1
